[refactor] liveshopping autocomplete 컴포넌트 분리
