+++ libs/components-admin/src/lib/kkshow-main/KkshowMainCarouselItemDialog.tsx
@@ -32,35 +32,9 @@ import {
 import React, { useMemo, useRef, useState } from 'react';
 import path from 'path';
 import { s3 } from '@project-lc/utils-s3';
-import { LiveShoppingWithGoods, useAdminLiveShoppingList } from '@project-lc/hooks';
-import { ChakraAutoComplete } from '@project-lc/components-core/ChakraAutoComplete';
+import { LiveShoppingWithGoods } from '@project-lc/hooks';
 import { Preview, readAsDataURL } from '@project-lc/components-core/ImageInputDialog';
-
-/** 크크쇼메인데이터 중 라이브방송정보(라이브방송,판매상품,방송인) 조회하기 위한 autocomplete컴포넌트 */
-export function LiveShoppingListAutoComplete({
-  onChange,
-}: {
-  onChange: (item: LiveShoppingWithGoods | null) => void;
-}): JSX.Element {
-  const { data: liveShoppingList } = useAdminLiveShoppingList({
-    enabled: true,
-  });
-  return (
-    <ChakraAutoComplete
-      options={
-        liveShoppingList
-          ? liveShoppingList.filter((l) =>
-              ['adjusting', 'confirmed'].includes(l.progress),
-            )
-          : []
-      }
-      getOptionLabel={(option) => option?.liveShoppingName || ''}
-      onChange={(newItem) => {
-        onChange(newItem || null);
-      }}
-    />
-  );
-}
+import { LiveShoppingListAutoComplete } from './LiveShoppingListAutoComplete';
 
 export function KkshowMainCarouselItemDialog({
   isOpen,
